Basic neorg help-page
This help-page is essentially a copy of the `EXAMPLE.norg` file.
However, it is accompanied by a `tags` file which vim understands and
parses to extract help tags.
In the future, an automatic generation of this tags file will be added.
For now, we have to update it manually. But the flex we can achieve with
having norg-formatted help pages is definitely worth that! 💪