WIP: TS-based concealing
Using the workaround explained in [1] we can conceal basic markup
entirely based on the TS syntax tree. While this has some minor edge
cases (especially considering the display of a colorcolumn).

This commit also introduces a new conceal setting called `volatile`
which will make icons.. well.. volatile! That means, that the concealing
will be lifted when the cursor is above the extmarks start position
making editing possible in a much easier fashion.

[1]: https://github.com/neovim/neovim/issues/15706#issuecomment-922253522


Former-commit-id: 73a1c9b0ca734729d61309394c5e2c39f3c69699