fix: jump to next/prev treesitter match on the same line
The `goto_next_query_match` and `goto_previous_query_match` functions in
the treesitter module are used to find the next and previous nodes to
jump to for a given treesitter query.

This is fine when used for headings, as a single line can only contain 1
heading. However, with the recent addition of jumping to links, this is
no longer correct as a line may contain multiple links.

The current behaviour would always take you to the first node of the
next line, even if there were matching nodes following your cursor on
the current line. This change fixes that by taking the column into
account when the matching node is the on the current line.