fix: do not skip anchors when jumping to next/prev link
Anchor definitions and anchor declarations were skipped when jumping to
the next or previous link. This was a bit surprising as you could jump
to links with a custom text for example. Since anchors serve a similar
purpose as links, it makes sense to be able to jump to them.