Fix TODO completion
Added the "(" as a trigger character, prevent ( from spawning a ) with
the updated regex and updated the completion table.

I'm not super familiar with the new syntax yet.. or the syntax in
general; so I'll probably end up reading up on the documentation to see
what other completions can be disregarded and what completions should be
added. But I'll do that in a future pull request.