Refactor: Moved `neorg` into its own module
Right now, file sequencing is still required, since all files treat the
Neorg module the same way as if it was a global variable, the only
difference being that every module imports it as a local at the top of
the file. This will require some refactoring down the line regarding
the sequencing of require-statements within a module.