Refactor: Moved `modules` and `events` to `modules/init.lua`
This patch is gargantuan because `modules` and `events` are two modules
that know each other's details very intimately, so the first solution
step is to move both of them into the same file to avoid elliptic
dependencies and the see how to proceed forward with their separation of
responsibilities. Right now it works(TM) but the `modules` init file is
a hot mess of functions from both `modules` and `events`.