Refactor: Greatly simplified `events` module and moved it into `core`
A lot to unpack here! First of all, the `event.base` class has been
redesigned to only contain a "quick event", which is an event that is
not aware of the user's context. The `event` class derives from
`event.base` and has dropped the `type` and `split_type` fields in
favour of `referrer.name` for `split_type[1]` and `name` for
`split_type[2]`, so event names have been dramatically shortened since
they now don't require their prefix up until the end of the substring
".events."  The field `content` has been renamed to `payload` because it
was called like that in some modules and sounds better to me, but that
can be reverted. The field `broadcast` has been dropped becuase it was
never used and is now a member function along with `send_to`, used to
send the events after creation in a single statement.

The event definition function has been deprecated because its return
value was not used. I will probably write a quick event definition class
that stores the command to run when calling the event.

A `types.lua` file has also been added for global types.