Update logic for adding li tags
This is done so that it can correctly render mulit-line list items. The
only real change to logic is that nested lists will be inlcuded inside
of a given list item (this may or may not be the correct way to handle
nested lists). My limited testing has resulted in HTML that rendered
correctly.