docs(README): lazy load neorg for lazy.nvim
Forcing the load of the neorg plugin with the lazy = false
statement leads to a 60ms increase in startup time, even when only
opening regular files and not using the plugin.

The main goal of lazy.nvim is to load plugins just-in-time and avoid
unecessary loads when not required.

Update the installation instructions for lazy.nvim users to lazy load
neorg upon `:Neorg` invocation, or when opening `norg` filetype files.